Add watermark feature
When the base code is distributed to students, each student
should download the files with their own personal watermarks
embedded in the files. There is a JavaScript script for
automatically adding those watermarks when a student downloads
base code files in the A+ course. Radar can then compare
students' submissions and check if multiple students submit files
with the same watermarks.

The watermarks are invisible (not printed) strings of whitespace
characters.