Clean up migrations for the watermark feature
Change the watermark model field to CharField. TextField does not
seem sensible since the watermark should have a max length.
However, using the max_length parameter in TextFields does not
set any limit to the database; it is only used in the HTML widget
for validation.

Update the framework versions in the README and upgrade
the requests library because it has bug fixes.