Update the Scalariform library for tokenizing Scala submissions
The new version of Scalariform is 0.2.10. It is the last release
from 2019. The library is no longer maintained.

The built package scalariform.jar now includes the Scalariform library
as well as the small CLI script ScalariformTokens.
The JAR package is built for Scala 2.13 and can not be run with older
Scala versions.

Add sbt config files for building the scalariform.jar package with sbt.
The command `sbt assembly` builds the JAR package with the dependencies.
The build output is written to the scalariform/target/scala-2.13
directory and it must be copied from there manually.

If there is a need to move on from the dead Scalariform library,
one alternative could be Scalafmt:
https://scalameta.org/scalafmt/