Switch the Memcached backend to pylibmc
The old backend python-memcached is deprecated and unmaintained.

Pylibmc is a Python client for Memcached written in the C language.
It uses libmemcached.

Note that Radar depends on Celery and Celery does not support
the Pymemcache client for Memcached at the time of writing this.
Both Django and Celery support pylibmc.

Open issue at Celery:
Add pymemcache support to the cache result backend
https://github.com/celery/celery/issues/5531

https://docs.celeryq.dev/en/stable/userguide/configuration.html#conf-cache-result-backend